library: fix compiler warning when wchar_t is 16-bit
Some architectures have wchar_t as 16-bit. So if you pass it
through printf() like functions, they are being promoted to
integer. However, the code to build Sys-T payload feeds
wchar_t to va_arg() which expects argument to be at least
the size of an integer, hence the warning:

  library/src/mipi_syst_api.c:947:34: warning: 'wchar_t' {aka 'short unsigned int'} is promoted to 'int' when passed through '...'

So fix it by promoting wchar_t to integer if needed.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>